requirements.txt
Flask==2.2.2
mysql-connector-python==8.0.32
PyYAML==6.0
requests==2.25.1
